requirements.txt
requests
anthropic 
python-dotenv
openmeteo-requests
pydantic-core
pandas
requests-cache
retry-requests
aiohttp
claude-agent-sdk
asyncio